move navigation file to its own package and move my_app to its associated state and notifier classes
